requirements.txt
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
xgboost>=2.0.0
ta>=0.11.0
textblob>=0.18.0
feedparser>=6.0.10
python-telegram-bot>=20.5
apscheduler>=3.10.4
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
joblib>=1.3.0
twilio>=8.10.0
pytrends>=4.9.0
beautifulsoup4>=4.12.0
customtkinter>=5.2.0
ntscraper>=0.3.0
GoogleNews>=1.6.0
certifi
pip-system-certs
